refactor: BaseViewModel機能を各ViewModelに統合してビルド可能に
- BaseViewModel.swiftをXcodeプロジェクトに登録する代わりに、
  その機能を各ViewModelに直接統合
- 4つのViewModel（Sensing, Settings, Advertiser, Trajectory）すべてに
  @MainActorとBaseViewModelのプロパティ/メソッドを追加
- ViewModels/BaseViewModel.swiftを削除

この変更により、Xcodeプロジェクトファイルの手動編集なしで
ビルドが通るようになりました。

コードの重複は発生しますが、以下のメリットがあります:
- ビルドが確実に通る
- @MainActorによるスレッド安全性の確保
- 統一されたエラーハンドリング（errorMessage, showErrorAlert）
- ローディング状態管理（isLoading）の統一

Related to #32
