+++ UWBViewerSystem/Presentation/Scenes/SensingTab/TrajectoryViewPage/TrajectoryViewModel.swift
@@ -9,7 +9,12 @@ import SwiftUI
 #endif
 
 @MainActor
-class TrajectoryViewModel: BaseViewModel {
+class TrajectoryViewModel: ObservableObject {
+    // BaseViewModel properties
+    @Published var errorMessage: String = ""
+    @Published var showErrorAlert: Bool = false
+    @Published var isLoading: Bool = false
+
     @Published var availableSessions: [SensingSession] = []
     @Published var selectedSession: SensingSession?
     @Published var trajectoryPoints: [TrajectoryPoint] = []
